# specs/002-todo-fullstack-webapp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Full-Stack Web Application API
  description: |
    REST API for Phase II Todo application.

    Authentication uses JWT tokens stored in httpOnly cookies.
    All task endpoints require authentication.
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: User registration, login, and logout
  - name: Tasks
    description: Task CRUD operations (requires authentication)

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            example:
              email: "user@example.com"
              password: "securepassword123"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Password must be at least 8 characters"
                code: "VALIDATION_ERROR"
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Email already exists"
                code: "CONFLICT"

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and set JWT cookie
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: "user@example.com"
              password: "securepassword123"
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: JWT token in httpOnly cookie
              schema:
                type: string
                example: "access_token=eyJ...; HttpOnly; Secure; SameSite=Lax; Path=/"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid email or password"
                code: "UNAUTHORIZED"

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Clear JWT cookie and end session
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              description: Clear JWT cookie
              schema:
                type: string
                example: "access_token=; HttpOnly; Secure; SameSite=Lax; Path=/; Max-Age=0"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"

  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: Get all tasks for the authenticated user
      operationId: listTasks
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Create a new task for the authenticated user
      operationId: createTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            example:
              title: "Buy groceries"
              description: "Milk, bread, eggs"
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task title is required"
                code: "VALIDATION_ERROR"
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    get:
      tags:
        - Tasks
      summary: Get a task
      description: Get a specific task by ID (must be owned by user)
      operationId: getTask
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: Update task title and/or description (must be owned by user)
      operationId: updateTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Buy groceries (updated)"
              description: "Milk, bread, eggs, cheese"
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Delete a task by ID (must be owned by user)
      operationId: deleteTask
      security:
        - cookieAuth: []
      responses:
        '204':
          description: Task deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/tasks/{taskId}/toggle:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    patch:
      tags:
        - Tasks
      summary: Toggle task completion
      description: Toggle the completion status of a task (must be owned by user)
      operationId: toggleTask
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Task toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT token stored in httpOnly cookie

  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password (minimum 8 characters)

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        email:
          type: string
          format: email
          description: User email address
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title (required)
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Task description (optional)

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Task description

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          description: Task ID
        title:
          type: string
          description: Task title
        description:
          type: string
          nullable: true
          description: Task description
        is_completed:
          type: boolean
          description: Completion status
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    ErrorResponse:
      type: object
      required:
        - detail
        - code
      properties:
        detail:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
          description: Error code

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Please log in to continue"
            code: "UNAUTHORIZED"

    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "You don't have permission to access this resource"
            code: "FORBIDDEN"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"
            code: "NOT_FOUND"